ShopCart/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cron-descriptor==1.4.5
cryptography==45.0.5
cssselect2==0.8.0
defusedxml==0.7.1
deprecation==2.1.0
Django==5.2.3
django-celery==3.1.17
django-celery-beat==2.8.1
django-crispy-forms==2.4
django-email-verification==0.3.3
django-environ==0.12.0
django-google-fonts==0.0.3
django-htmx==1.23.2
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
Faker==37.4.0
flower==2.0.1
fonttools==4.58.5
gunicorn==21.2.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stripe==12.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
